Use variable test keys when running CI
We currently use the in-repository keys for testing in CI. This has the
downside that once they expire, we can't re-run these tests inside CI.

Use keys that are passed as environment variables, so once we
reconfigure CI to use new keys, this also affects re-runs of older
builds.

The test keys are still kept in-repo, so as not to complicate testing
for developers and contributors.