Check that a responses are not errors
AFIP allows us to create valid tickets with invalid key/CUIT pairs, so
we can end up with tickets that fail on any service.

Due to how suds works, we can't quite catch these sort of errors on some
middleware level.

While this demands for a larger refactor as to how we do WS calls, this
hack should at suffice for now.