Drop ReceiptBatch
ReceiptBatch was an attempt to directly reflect the AFIP's API on this
library, but it's existence was rather confusing, complex, and added a
lot of burden both to the devs and downstream users.

Completely drop the concept of a materialized batch; expose a simpler
(more Pythonic, or Djanoic) API to validate receipts, and handle the
ugly stuff behind the scenes.