Remove no-longer necessary monkey-patching
This used to be necessary a couple of years ago -- either the migration
to zeep fixed the underlying issue for us, or AFIP fixed the issue on
their side (most likely the latter).  The big upside is that no monkey
patching is necessary, so this should greatly improve compatibility with
odd pythons.