Simplify TaxPayerExtras into TaxPayer
The original intent of this library was to provide only receipt validation,
when printable were later implemented, they were designed as an "optional"
extra.

Because of this, the fields required for printing [but not validating] were
kept in different tables. This has been very confusing for users. I've gotten
lots of emails, and, ultimately, it seems it was bad design, since everyone
also wants to use this library to generate the PDFs too.

This is part of the work necessary to improve on this design.