Use natural keys for serialising auth tickets
When running live tests locally, an imported ticket specifies an
explicit pk. After importing this, postgres' internal sequence is not
updated, so the next-created ticket has the same pk and the creation
fails due to an integrity error.

Use a natural key for serialising auth tickets, so that importing one
will use a sequence-generated pk, and increase postgres's internal
counter.

This should also make it safe to export and import auth tickets in
production systems in general (assuming the id of related tables match
on both systems).