Switch to sphinxcontrib-django
This project has a helper module which improves the generated
documentation for Django models and other types. Back in 2016 when I
wrote it, there was no library to do this.

Nowadays, `sphinxcontrib-django` does the same thing but better and
handles a lot of extra details that my code doesn't.

Use the external library, and drop the in-tree implementation.